Get relevant parts of the powerset only.
This makes sense since if there is a zero sum with length > len(summans) / 2 there is a shorter one.

This makes the code twice as slow, because adding these longer sequences is beneficial for the memoization on later sequences.

To deal with that - might as well add all permutations of a sequence when we add the sequence. The function that does that is called "get similar sums" for now.