Create a sumdigraph class
the idea is that it would be enough to traverse from a zero at the bottom layer and try to find a zero on the way. I'm pretty sure the traversal is as expensive as the current subset method, but - the sum is calculated once! and currently the main bottleneck is the sum calculation.